build: separate module for frontend
This allows us to simplify the webpack configuration and the gradle
build scripts.
