+++ subprojects/frontend/src/editor/EditorArea.tsx
