+++ subprojects/frontend/src/editor/findOccurrences.ts
