+++ subprojects/frontend/src/editor/semanticHighlighting.ts
