+++ subprojects/frontend/src/global.d.ts
