+++ subprojects/frontend/tsconfig.json
@@ -9,9 +9,10 @@
     "strict": true,
     "noImplicitOverride": true,
     "noImplicitReturns": true,
+    "exactOptionalPropertyTypes": false,
     "noEmit": true,
     "skipLibCheck": true
   },
-  "include": ["./src/main/js/**/*"],
-  "exclude": ["./src/main/js/xtext/**/*"]
+  "include": ["./src/**/*"],
+  "exclude": ["./build/generated/sources/lezer/*"]
 }
