+++ subprojects/frontend/tsconfig.sonar.json
@@ -9,10 +9,8 @@
     "strict": true,
     "noImplicitOverride": true,
     "noImplicitReturns": true,
-    "exactOptionalPropertyTypes": false,
     "noEmit": true,
     "skipLibCheck": true
   },
-  "include": ["./src/main/js/**/*"],
-  "exclude": ["./build/generated/sources/lezer/*"]
+  "include": ["./src/**/*"]
 }
